requirements.txt
tensorflow == 2.2.0
tensorflow-addons == 0.10.0
tensorflow-datasets == 2.0.0
absl-py >= 0.6.1
h5py >= 2.10.0
matplotlib >= 2.2.5
numpy >= 1.15.4
psutil >= 5.7.0
scipy >= 1.1.0
tqdm >= 4.45.0
OpenEXR >= 1.3.2
termcolor >= 1.1.0
trimesh >= 2.37.22
# Required by trimesh.
networkx
